G2_24 G2_32 first logic to update

+++ demo/src/main/java/com/architec/demo/repositories/AssignmentRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
 public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
     List<Assignment> findByMaxStudents(int maxStudents);
     List<Assignment> findByAssignerUserId(long assignerUserId);
-    List<Assignment> findByAssignmentId(int assignmentId);
+    List<Assignment> findByAssignmentId(long assignmentId);
 }
